+++ src/UI/Props/Text/FontSize.cs
@@ -35,9 +35,6 @@ namespace WForest.UI.Props.Text
             {
                 if (text.FontSize == _size) return;
                 text.FontSize = _size >= 0 ? _size : throw new ArgumentException("FontSize cannot be negative.");
-                var (x, y, _, _) = text.Space;
-                var (w, h) = text.Font.MeasureText(text.TextString, _size);
-                WidgetsSpaceHelper.UpdateSpace(widget, new Rectangle(x, y, (int) w, (int) h));
             }
             else
             {